requirement.txt
streamlit
opencv-python==4.5.3.56
pillow
numpy
